Remove the support of skip_tokens in llm_attr target
Summary:
The `skip_token` was originally introduced in D62775617 . this can be further improved.

This diff reverts it.

Transformer is auto-regression model, so the sequence of token cannot be changed. It doesn't make sense to support filter tokens within the sequence.

The correct behavior is to only trim leading extra special tokens (if any) when `target` is tokenized alone (w/o input)

Differential Revision: D89682911