Add format_pixel_attr in ImageMaskInput (#1739)
Summary:

Add a new function `format_pixel_attr` in `ImageMaskInput` to hold the orginal logic of `format_attr`. This fun project the interpretable feature attribution to the image size, which will be useful to render any image heatmap.

Revise `format_attr` to be dummy fun keeping the shape of interpretable feature attribution, since this is required for now to be compatible with `LLMAttributionResult`

Differential Revision: D89908572